// src/app/shared/character.interface.ts
export interface CharacterInterface {
  [x: string]: any;
  id: number;
  name: string;
  height: number;
  mass: number;
  gender: string;
  homeworld: string;
  wiki: string;
  image: string;
  born: number;
  bornLocation: string;
  died: number;
  diedLocation: string;
  species: string;
  hairColor: string;
  eyeColor: string;
  skinColor: string;
  cybernetics: string;
  affiliations: [];
  masters: [];
  apprentices: [];
  formerAffiliations: [];
}